+++ packages/vite-plugin-istanbul-widget/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vite-plugin-istanbul-widget
 
+## 1.5.2
+
+### Patch Changes
+
+- update istanbul-widget import path
+- Updated dependencies
+  - istanbul-widget@1.5.1
+
 ## 1.5.1
 
 ### Patch Changes